Added empty project and initial process writing
Added an empty project that doesn't yet have the TANKS! project in it because I need to think a bit more about how to strip that back.

Added initial process writing that represents the beginning of trying to figure out what this project is about. In particular, struggling with how to frame the game beyond "Oh, it's something about using all the possibilities of Unity one by one to affect a specific pre-existing game". That sounds "formal" and therefore doable, but doesn't actually make so much sense from a design perspective. Much agonising ensues.