Added completed TANKS! tutorial build, decision about LightingData
This included making changes to the editor settings to do the visible meta files and forcing text serialization for (hopeful) compatibility with Git as well as looking into the idea of using GitLFS and deciding against it (I'm hoping this project isn't going to be a behemoth).

This also included pulling down the the tutorial asset pack and deleting all the "not done yet" bits in favour of just the fully completed game. Also a brief moment of confusion over pre-computed realtime GI where I deleted the LightingData for the scene thinking it was unnecessary for realtime lighting, but then finding out that it is necessary for the precomputed stuff, so regenerated it. (Will need to remember, presumably to generate this for each scene, which could balloon the game in an ugly way? Would it be better to go totally realtime with no precomputation? (It was a bit darker but otherwise totally fine visually.)

This is a good example of having to make a technical decision early that I'm not really very well informed about. I think the correct answer is to _delete_ the precomputes and only add them in at the end if they seem needed. I'll do that now. Done. So the game looks a little different from the "real" version, but I can easily regenerate the lighting at the end if I want to, and this way I know it's pure realtime for the moment. Yeah it's fine. And at least for the moment I avoid some major-seeming build sizes? Though it's a question of course whether those LightingData files are actually reflected 1:1 in the final build?

Anyway, this is what we're commiting for the moment. 80MB on disk, probably closer to 50MB on GitHub by omitting things like Library and Temp etc.