Light Experiments: Added a version of the game where the individual tanks have headlights (two spotlights per tank) that can be turned on or off. The scene has no lighting at all. The spotlights cast shadows.
? This was hugely an exercise in early frustrations with the technology, which was not unexpected but sucked anyway. In particular it took a long while to comprehend various bugs and issue especially with getting the code to play nice. There are namespace issues involved in writing different versions of the script, for example. One way to "solve" this would be to have every variation in a branch, but we need to bear in mind that the final result has to be integrated, meaning many different scripts must coexist in the same project. Fortunately the basic structure of the game is pretty simple, so this should be doable, but it's a major issue with the project (and will potentially hinder students).

My solution to this for now is going to be to make all the original scripts sacrosanct (no editing them) but to have them all have inheritable methods (e.g. all their scripts will have their methods marked "protected virtual" so that I can inherit from them and override them (protected override). This way the core behaviour is the same and I can graft on new code in a subclass that's separate (and has a name relevant to the project in question. So for example I created a "HeadlightsTankMovement.cs" in this case which is a subclass of TankMovement and adds on a check for toggling the headlights on and off.

That works. I guess there will almost certainly be cases where I need to do related things for the entire GameManager too, and that will still be fine I guess.  (The game manager is just an Empty GameObject witht he script on it and an audio source that plays the music.) So that will (for now!) be the solution. One "nice" thing is the degree to which it maintains the original structure of the game and positions whatever I do as separate from it, which has a degree of rigour I suppose I appreciate? I'm a little afraid of the instant spike in complexity required simply to write code? But maybe from here it will be "smooth sailing". Ha ha. No.

+ This version also required what I imagine will be a common practice which was creating a new Prefab of the tank specifically for my new version. In this case that meant duplicating the existing prefab and adding the headlights to it, and then adding the new movement script. That actually worked pretty seamlessly I'd have to say, thank god.

+ This version also required editing the input. Note that that edits the input for the ENTIRE PROJECT so that's worth thinking about whether there might be problems if I wanted yet more controls? I added Alt1 and Alt2 to the control scheme to be able to toggle the lights in this case.

? Overall this work session did at least get a variation of the game working, albeit a very simple one. I ran into some key technical difficulties and surmounted them (for the moment). I also had the chance to witness a couple of actual Light Properties. Most notably when you're driving around with headlights on it's surprisingly easy to navigate. And it makes the environment much more mysterious and kind of interesting actually. Also makes it much weirder that the shells have point lights on them - feels like it breaks 'realism' to some extent. Also very nice to see the shadows in realtime... created some good noir/nighttime feeling moving shadows just be driving around. A staircase cast great shadows. Felt very expressive visually (but then also kind of dead because of the nature of the game itself?)

? I'm not sure how this has made me feel about the 'rigourous path' I wrote about in my process journal today. If I only change the lights, it currently feels like I might end up with kind of empty modifications? Maybe I need to keep playing with it though. Almost certainly yes. Also as it becomes more fluid to edit I should hopefully get a better sense of the "language" involved in this. The coversation with materials presumably gets more sophisticated as you learn to speak to them more subtly? We'll see on that.